Split shared/curriculum.py: separate the manager from the SB3 callbacks
curriculum.py was 1,316 lines holding a backend-independent manager plus
nine SB3 callbacks behind a shared optional-import shim. Split it so the
manager no longer sits in the same module as the SB3 integration:

  sb3_compat.py     the optional-SB3 import shim
  manager.py        StageThreshold + CurriculumManager (no SB3 dependency)
  schedules.py      constant schedules, entropy-coefficient decay
  advancement.py    stage entry, reward ramping, gate evaluation
  early_stopping.py abort a stage whose eval return has collapsed
  checkpoints.py    checkpoint selection/publication, VecNormalize stats

The package __init__ re-exports the same public surface, so callers and
train_base's lazy imports are unchanged.

_SB3_AVAILABLE is read off the shim module rather than copied into each
callback module, so there is one patch point for it instead of one per
module; test_curriculum.py patches that, and the two metrics helpers at
the module they are used from.
